+++ sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/audio/converters/SimpleRateConverter.hpp
@@ -1,0 +1,106 @@
+#pragma once
+
+#include <array>
+#include <HyperSonicDrivers/audio/scummvm/RateConverter.hpp>
+#include <HyperSonicDrivers/audio/scummvm/AudioStream.hpp>
+
+#include <SDL2/SDL_log.h>
+
+namespace HyperSonicDrivers::audio::converters
+{
+    /**
+         * Audio rate converter based on simple resampling. Used when no
+         * interpolation is required.
+         *
+         * Limited to sampling frequency <= 65535 Hz.
+         */
+    template<bool stereo, bool reverseStereo>
+    class SimpleRateConverter : public scummvm::RateConverter {
+    protected:
+        std::array<int16_t, scummvm::INTERMEDIATE_BUFFER_SIZE> inBuf = {};
+        const int16_t* inPtr = nullptr;
+        int inLen = 0;
+
+        /** position of how far output is ahead of input */
+        /** Holds what would have been opos-ipos */
+        long opos = 0;
+
+        /** fractional position increment in the output stream */
+        long opos_inc = 0;
+
+    public:
+        SimpleRateConverter(uint32_t inrate, uint32_t outrate);
+        int flow(scummvm::AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r);
+        int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
+            return scummvm::ST_SUCCESS;
+        }
+    };
+
+    /*
+         * Prepare processing.
+         */
+    template<bool stereo, bool reverseStereo>
+    SimpleRateConverter<stereo, reverseStereo>::SimpleRateConverter(uint32_t inrate, uint32_t outrate) {
+        if ((inrate % outrate) != 0) {
+            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "Input rate must be a multiple of output rate to use rate effect");
+        }
+
+        if (inrate >= 65536 || outrate >= 65536) {
+            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "rate effect can only handle rates < 65536");
+        }
+
+        opos = 1;
+
+        /* increment */
+        opos_inc = inrate / outrate;
+
+        inLen = 0;
+    }
+
+    /*
+     * Processed signed long samples from ibuf to obuf.
+     * Return number of sample pairs processed.
+     */
+    template<bool stereo, bool reverseStereo>
+    int SimpleRateConverter<stereo, reverseStereo>::flow(scummvm::AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
+        int16_t* ostart, * oend;
+
+        ostart = obuf;
+        oend = obuf + osamp * 2;
+
+        while (obuf < oend) {
+
+            // read enough input samples so that opos >= 0
+            do {
+                // Check if we have to refill the buffer
+                if (inLen == 0) {
+                    inPtr = inBuf.data();
+                    inLen = input.readBuffer(inBuf.data(), inBuf.size());
+                    if (inLen <= 0)
+                        return (obuf - ostart) / 2;
+                }
+                inLen -= (stereo ? 2 : 1);
+                opos--;
+                if (opos >= 0) {
+                    inPtr += (stereo ? 2 : 1);
+                }
+            } while (opos >= 0);
+
+            int16_t out0, out1;
+            out0 = *inPtr++;
+            out1 = (stereo ? *inPtr++ : out0);
+
+            // Increment output position
+            opos += opos_inc;
+
+            // output left channel
+            scummvm::clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / scummvm::Mixer::MaxVolume::MIXER);
+
+            // output right channel
+            scummvm::clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / scummvm::Mixer::MaxVolume::MIXER);
+
+            obuf += 2;
+        }
+        return (obuf - ostart) / 2;
+    }
+}
