+++ sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/audio/scummvm/RateConverter.cpp
@@ -5,7 +5,13 @@
 #include <HyperSonicDrivers/audio/scummvm/RateConverter.hpp>
 #include <HyperSonicDrivers/audio/scummvm/AudioStream.hpp>
 #include <HyperSonicDrivers/audio/scummvm/Mixer.hpp>
+
+#include <HyperSonicDrivers/audio/converters/SimpleRateConverter.hpp>
+#include <HyperSonicDrivers/audio/converters/LinearRateConverter.hpp>
+#include <HyperSonicDrivers/audio/converters/CopyRateConverter.hpp>
+
 #include <SDL2/SDL_log.h>
+
 
 namespace HyperSonicDrivers::audio::scummvm
 {
@@ -21,23 +27,23 @@ namespace HyperSonicDrivers::audio::scummvm
       * @{
       */
 
-      /**
-       * The precision of the fractional (fixed-point) type that is defined below.
-       * Normally, you should never need to modify this value.
-       */
-    enum {
-        FRAC_BITS = 16,
-        FRAC_LO_MASK = ((1L << FRAC_BITS) - 1),
-        FRAC_HI_MASK = ((1L << FRAC_BITS) - 1) << FRAC_BITS,
+    //  /**
+    //   * The precision of the fractional (fixed-point) type that is defined below.
+    //   * Normally, you should never need to modify this value.
+    //   */
+    //enum {
+    //    FRAC_BITS = 16,
+    //    FRAC_LO_MASK = ((1L << FRAC_BITS) - 1),
+    //    FRAC_HI_MASK = ((1L << FRAC_BITS) - 1) << FRAC_BITS,
 
-        FRAC_ONE = (1L << FRAC_BITS),		// 1.0
-        FRAC_HALF = (1L << (FRAC_BITS - 1))	// 0.5
-    };
+    //    FRAC_ONE = (1L << FRAC_BITS),		// 1.0
+    //    FRAC_HALF = (1L << (FRAC_BITS - 1))	// 0.5
+    //};
 
-    /**
-     * Fixed-point fractions, used by the sound rate converter and other code.
-     */
-    typedef int32_t frac_t;
+    ///**
+    // * Fixed-point fractions, used by the sound rate converter and other code.
+    // */
+    //typedef int32_t frac_t;
 
     inline frac_t doubleToFrac(double value) { return (frac_t)(value * static_cast<int>(FRAC_ONE)); }
     inline double fracToDouble(frac_t value) { return ((double)value) / static_cast<int>(FRAC_ONE); }
@@ -64,115 +70,115 @@ namespace HyperSonicDrivers::audio::scummvm
     * target processor and various other factors), at which it will decrease
     * again.
     */
-#define INTERMEDIATE_BUFFER_SIZE 512
+//#define INTERMEDIATE_BUFFER_SIZE 512
 
-    /**
-    * The default fractional type in frac.h (with 16 fractional bits) limits
-    * the rate conversion code to 65536Hz audio: we need to able to handle
-    * 96kHz audio, so we use fewer fractional bits in this code.
-    */
-    enum {
-        FRAC_BITS_LOW = 15,
-        FRAC_ONE_LOW = (1L << FRAC_BITS_LOW),
-        FRAC_HALF_LOW = (1L << (FRAC_BITS_LOW - 1))
-    };
+    ///**
+    //* The default fractional type in frac.h (with 16 fractional bits) limits
+    //* the rate conversion code to 65536Hz audio: we need to able to handle
+    //* 96kHz audio, so we use fewer fractional bits in this code.
+    //*/
+    //enum {
+    //    FRAC_BITS_LOW = 15,
+    //    FRAC_ONE_LOW = (1L << FRAC_BITS_LOW),
+    //    FRAC_HALF_LOW = (1L << (FRAC_BITS_LOW - 1))
+    //};
 
-    /**
-     * Audio rate converter based on simple resampling. Used when no
-     * interpolation is required.
-     *
-     * Limited to sampling frequency <= 65535 Hz.
-     */
-    template<bool stereo, bool reverseStereo>
-    class SimpleRateConverter : public RateConverter {
-    protected:
-        std::array<int16_t, INTERMEDIATE_BUFFER_SIZE> inBuf = {};
-        const int16_t* inPtr = nullptr;
-        int inLen = 0;
+    ///**
+    // * Audio rate converter based on simple resampling. Used when no
+    // * interpolation is required.
+    // *
+    // * Limited to sampling frequency <= 65535 Hz.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //class SimpleRateConverter : public RateConverter {
+    //protected:
+    //    std::array<int16_t, INTERMEDIATE_BUFFER_SIZE> inBuf = {};
+    //    const int16_t* inPtr = nullptr;
+    //    int inLen = 0;
 
-        /** position of how far output is ahead of input */
-        /** Holds what would have been opos-ipos */
-        long opos = 0;
+    //    /** position of how far output is ahead of input */
+    //    /** Holds what would have been opos-ipos */
+    //    long opos = 0;
 
-        /** fractional position increment in the output stream */
-        long opos_inc = 0;
+    //    /** fractional position increment in the output stream */
+    //    long opos_inc = 0;
 
-    public:
-        SimpleRateConverter(uint32_t inrate, uint32_t outrate);
-        int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r);
-        int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
-            return ST_SUCCESS;
-        }
-    };
+    //public:
+    //    SimpleRateConverter(uint32_t inrate, uint32_t outrate);
+    //    int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r);
+    //    int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
+    //        return ST_SUCCESS;
+    //    }
+    //};
 
 
-    /*
-     * Prepare processing.
-     */
-    template<bool stereo, bool reverseStereo>
-    SimpleRateConverter<stereo, reverseStereo>::SimpleRateConverter(uint32_t inrate, uint32_t outrate) {
-        if ((inrate % outrate) != 0) {
-            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "Input rate must be a multiple of output rate to use rate effect");
-        }
+    ///*
+    // * Prepare processing.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //SimpleRateConverter<stereo, reverseStereo>::SimpleRateConverter(uint32_t inrate, uint32_t outrate) {
+    //    if ((inrate % outrate) != 0) {
+    //        SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "Input rate must be a multiple of output rate to use rate effect");
+    //    }
 
-        if (inrate >= 65536 || outrate >= 65536) {
-            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "rate effect can only handle rates < 65536");
-        }
+    //    if (inrate >= 65536 || outrate >= 65536) {
+    //        SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "rate effect can only handle rates < 65536");
+    //    }
 
-        opos = 1;
+    //    opos = 1;
 
-        /* increment */
-        opos_inc = inrate / outrate;
+    //    /* increment */
+    //    opos_inc = inrate / outrate;
 
-        inLen = 0;
-    }
+    //    inLen = 0;
+    //}
 
-    /*
-     * Processed signed long samples from ibuf to obuf.
-     * Return number of sample pairs processed.
-     */
-    template<bool stereo, bool reverseStereo>
-    int SimpleRateConverter<stereo, reverseStereo>::flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
-        int16_t* ostart, * oend;
+    ///*
+    // * Processed signed long samples from ibuf to obuf.
+    // * Return number of sample pairs processed.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //int SimpleRateConverter<stereo, reverseStereo>::flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
+    //    int16_t* ostart, * oend;
 
-        ostart = obuf;
-        oend = obuf + osamp * 2;
+    //    ostart = obuf;
+    //    oend = obuf + osamp * 2;
 
-        while (obuf < oend) {
+    //    while (obuf < oend) {
 
-            // read enough input samples so that opos >= 0
-            do {
-                // Check if we have to refill the buffer
-                if (inLen == 0) {
-                    inPtr = inBuf.data();
-                    inLen = input.readBuffer(inBuf.data(), inBuf.size());
-                    if (inLen <= 0)
-                        return (obuf - ostart) / 2;
-                }
-                inLen -= (stereo ? 2 : 1);
-                opos--;
-                if (opos >= 0) {
-                    inPtr += (stereo ? 2 : 1);
-                }
-            } while (opos >= 0);
+    //        // read enough input samples so that opos >= 0
+    //        do {
+    //            // Check if we have to refill the buffer
+    //            if (inLen == 0) {
+    //                inPtr = inBuf.data();
+    //                inLen = input.readBuffer(inBuf.data(), inBuf.size());
+    //                if (inLen <= 0)
+    //                    return (obuf - ostart) / 2;
+    //            }
+    //            inLen -= (stereo ? 2 : 1);
+    //            opos--;
+    //            if (opos >= 0) {
+    //                inPtr += (stereo ? 2 : 1);
+    //            }
+    //        } while (opos >= 0);
 
-            int16_t out0, out1;
-            out0 = *inPtr++;
-            out1 = (stereo ? *inPtr++ : out0);
+    //        int16_t out0, out1;
+    //        out0 = *inPtr++;
+    //        out1 = (stereo ? *inPtr++ : out0);
 
-            // Increment output position
-            opos += opos_inc;
+    //        // Increment output position
+    //        opos += opos_inc;
 
-            // output left channel
-            clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
+    //        // output left channel
+    //        clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
 
-            // output right channel
-            clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
+    //        // output right channel
+    //        clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
 
-            obuf += 2;
-        }
-        return (obuf - ostart) / 2;
-    }
+    //        obuf += 2;
+    //    }
+    //    return (obuf - ostart) / 2;
+    //}
 
     /**
      * Audio rate converter based on simple linear Interpolation.
@@ -185,191 +191,191 @@ namespace HyperSonicDrivers::audio::scummvm
      * Limited to sampling frequency <= 65535 Hz.
      */
 
-    template<bool stereo, bool reverseStereo>
-    class LinearRateConverter : public RateConverter {
-    protected:
-        std::array<int16_t, INTERMEDIATE_BUFFER_SIZE> inBuf = {};
-        const int16_t* inPtr = nullptr;
-        int inLen = 0;
+    //template<bool stereo, bool reverseStereo>
+    //class LinearRateConverter : public RateConverter {
+    //protected:
+    //    std::array<int16_t, INTERMEDIATE_BUFFER_SIZE> inBuf = {};
+    //    const int16_t* inPtr = nullptr;
+    //    int inLen = 0;
 
-        /** fractional position of the output stream in input stream unit */
-        frac_t opos = 0;
+    //    /** fractional position of the output stream in input stream unit */
+    //    frac_t opos = 0;
 
-        /** fractional position increment in the output stream */
-        frac_t opos_inc = 0;
+    //    /** fractional position increment in the output stream */
+    //    frac_t opos_inc = 0;
 
-        /** last sample(s) in the input stream (left/right channel) */
-        int16_t ilast0 = 0;
-        int16_t ilast1 = 0;
-        /** current sample(s) in the input stream (left/right channel) */
-        int16_t icur0 = 0;
-        int16_t icur1 = 0;
+    //    /** last sample(s) in the input stream (left/right channel) */
+    //    int16_t ilast0 = 0;
+    //    int16_t ilast1 = 0;
+    //    /** current sample(s) in the input stream (left/right channel) */
+    //    int16_t icur0 = 0;
+    //    int16_t icur1 = 0;
 
-    public:
-        LinearRateConverter(uint32_t inrate, uint32_t outrate);
-        int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r);
-        int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
-            return ST_SUCCESS;
-        }
-    };
+    //public:
+    //    LinearRateConverter(uint32_t inrate, uint32_t outrate);
+    //    int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r);
+    //    int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
+    //        return ST_SUCCESS;
+    //    }
+    //};
 
 
-    /*
-     * Prepare processing.
-     */
-    template<bool stereo, bool reverseStereo>
-    LinearRateConverter<stereo, reverseStereo>::LinearRateConverter(uint32_t inrate, uint32_t outrate) {
-        if (inrate >= 131072 || outrate >= 131072) {
-            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "rate effect can only handle rates < 131072");
-        }
+    ///*
+    // * Prepare processing.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //LinearRateConverter<stereo, reverseStereo>::LinearRateConverter(uint32_t inrate, uint32_t outrate) {
+    //    if (inrate >= 131072 || outrate >= 131072) {
+    //        SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "rate effect can only handle rates < 131072");
+    //    }
 
-        opos = FRAC_ONE_LOW;
+    //    opos = FRAC_ONE_LOW;
 
-        // Compute the linear interpolation increment.
-        // This will overflow if inrate >= 2^17, and underflow if outrate >= 2^17.
-        // Also, if the quotient of the two rate becomes too small / too big, that
-        // would cause problems, but since we rarely scale from 1 to 65536 Hz or vice
-        // versa, I think we can live with that limitation ;-).
-        opos_inc = (inrate << FRAC_BITS_LOW) / outrate;
+    //    // Compute the linear interpolation increment.
+    //    // This will overflow if inrate >= 2^17, and underflow if outrate >= 2^17.
+    //    // Also, if the quotient of the two rate becomes too small / too big, that
+    //    // would cause problems, but since we rarely scale from 1 to 65536 Hz or vice
+    //    // versa, I think we can live with that limitation ;-).
+    //    opos_inc = (inrate << FRAC_BITS_LOW) / outrate;
 
-        ilast0 = ilast1 = 0;
-        icur0 = icur1 = 0;
+    //    ilast0 = ilast1 = 0;
+    //    icur0 = icur1 = 0;
 
-        inLen = 0;
-    }
+    //    inLen = 0;
+    //}
 
-    /*
-     * Processed signed long samples from ibuf to obuf.
-     * Return number of sample pairs processed.
-     */
-    template<bool stereo, bool reverseStereo>
-    int LinearRateConverter<stereo, reverseStereo>::flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
-        int16_t* ostart, * oend;
+    ///*
+    // * Processed signed long samples from ibuf to obuf.
+    // * Return number of sample pairs processed.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //int LinearRateConverter<stereo, reverseStereo>::flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
+    //    int16_t* ostart, * oend;
 
-        ostart = obuf;
-        oend = obuf + osamp * 2;
+    //    ostart = obuf;
+    //    oend = obuf + osamp * 2;
 
-        while (obuf < oend) {
+    //    while (obuf < oend) {
 
-            // read enough input samples so that opos < 0
-            while ((frac_t)FRAC_ONE_LOW <= opos) {
-                // Check if we have to refill the buffer
-                if (inLen == 0) {
-                    inPtr = inBuf.data();
-                    inLen = input.readBuffer(inBuf.data(), inBuf.size());
-                    if (inLen <= 0)
-                        return (obuf - ostart) / 2;
-                }
-                inLen -= (stereo ? 2 : 1);
-                ilast0 = icur0;
-                icur0 = *inPtr++;
-                if (stereo) {
-                    ilast1 = icur1;
-                    icur1 = *inPtr++;
-                }
-                opos -= FRAC_ONE_LOW;
-            }
+    //        // read enough input samples so that opos < 0
+    //        while ((frac_t)FRAC_ONE_LOW <= opos) {
+    //            // Check if we have to refill the buffer
+    //            if (inLen == 0) {
+    //                inPtr = inBuf.data();
+    //                inLen = input.readBuffer(inBuf.data(), inBuf.size());
+    //                if (inLen <= 0)
+    //                    return (obuf - ostart) / 2;
+    //            }
+    //            inLen -= (stereo ? 2 : 1);
+    //            ilast0 = icur0;
+    //            icur0 = *inPtr++;
+    //            if (stereo) {
+    //                ilast1 = icur1;
+    //                icur1 = *inPtr++;
+    //            }
+    //            opos -= FRAC_ONE_LOW;
+    //        }
 
-            // Loop as long as the outpos trails behind, and as long as there is
-            // still space in the output buffer.
-            while (opos < (frac_t)FRAC_ONE_LOW && obuf < oend) {
-                // interpolate
-                int16_t out0, out1;
-                out0 = (int16_t)(ilast0 + (((icur0 - ilast0) * opos + FRAC_HALF_LOW) >> FRAC_BITS_LOW));
-                out1 = (stereo ?
-                    (int16_t)(ilast1 + (((icur1 - ilast1) * opos + FRAC_HALF_LOW) >> FRAC_BITS_LOW)) :
-                    out0);
+    //        // Loop as long as the outpos trails behind, and as long as there is
+    //        // still space in the output buffer.
+    //        while (opos < (frac_t)FRAC_ONE_LOW && obuf < oend) {
+    //            // interpolate
+    //            int16_t out0, out1;
+    //            out0 = (int16_t)(ilast0 + (((icur0 - ilast0) * opos + FRAC_HALF_LOW) >> FRAC_BITS_LOW));
+    //            out1 = (stereo ?
+    //                (int16_t)(ilast1 + (((icur1 - ilast1) * opos + FRAC_HALF_LOW) >> FRAC_BITS_LOW)) :
+    //                out0);
 
-                // output left channel
-                clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
+    //            // output left channel
+    //            clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
 
-                // output right channel
-                clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
+    //            // output right channel
+    //            clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
 
-                obuf += 2;
+    //            obuf += 2;
 
-                // Increment output position
-                opos += opos_inc;
-            }
-        }
-        return (obuf - ostart) / 2;
-    }
+    //            // Increment output position
+    //            opos += opos_inc;
+    //        }
+    //    }
+    //    return (obuf - ostart) / 2;
+    //}
 
-    /**
-     * Simple audio rate converter for the case that the inrate equals the outrate.
-     */
-    template<bool stereo, bool reverseStereo>
-    class CopyRateConverter : public RateConverter {
-        int16_t* _buffer;
-        uint32_t _bufferSize;
-    public:
-        CopyRateConverter() : _buffer(0), _bufferSize(0) {}
-        ~CopyRateConverter() {
-            free(_buffer);
-        }
+    ///**
+    // * Simple audio rate converter for the case that the inrate equals the outrate.
+    // */
+    //template<bool stereo, bool reverseStereo>
+    //class CopyRateConverter : public RateConverter {
+    //    int16_t* _buffer;
+    //    uint32_t _bufferSize;
+    //public:
+    //    CopyRateConverter() : _buffer(0), _bufferSize(0) {}
+    //    ~CopyRateConverter() {
+    //        free(_buffer);
+    //    }
 
-        virtual int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
-            assert(input.isStereo() == stereo);
+    //    virtual int flow(AudioStream& input, int16_t* obuf, uint32_t osamp, uint16_t vol_l, uint16_t vol_r) {
+    //        assert(input.isStereo() == stereo);
 
-            int16_t* ptr;
-            uint32_t len;
+    //        int16_t* ptr;
+    //        uint32_t len;
 
-            int16_t* ostart = obuf;
+    //        int16_t* ostart = obuf;
 
-            if (stereo)
-                osamp *= 2;
+    //        if (stereo)
+    //            osamp *= 2;
 
-            // Reallocate temp buffer, if necessary
-            if (osamp > _bufferSize) {
-                free(_buffer);
-                _buffer = (int16_t*)malloc(osamp * 2);
-                _bufferSize = osamp;
-            }
+    //        // Reallocate temp buffer, if necessary
+    //        if (osamp > _bufferSize) {
+    //            free(_buffer);
+    //            _buffer = (int16_t*)malloc(osamp * 2);
+    //            _bufferSize = osamp;
+    //        }
 
-            if (_buffer == nullptr) {
-                SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "[CopyRateConverter::flow] Cannot allocate memory for temp buffer");
-                return 0;
-            }
+    //        if (_buffer == nullptr) {
+    //            SDL_LogError(SDL_LOG_CATEGORY_AUDIO, "[CopyRateConverter::flow] Cannot allocate memory for temp buffer");
+    //            return 0;
+    //        }
 
-            // Read up to 'osamp' samples into our temporary buffer
-            len = input.readBuffer(_buffer, osamp);
+    //        // Read up to 'osamp' samples into our temporary buffer
+    //        len = input.readBuffer(_buffer, osamp);
 
-            // Mix the data into the output buffer
-            ptr = _buffer;
-            for (; len > 0; len -= (stereo ? 2 : 1)) {
-                int16_t out0, out1;
-                out0 = *ptr++;
-                out1 = (stereo ? *ptr++ : out0);
+    //        // Mix the data into the output buffer
+    //        ptr = _buffer;
+    //        for (; len > 0; len -= (stereo ? 2 : 1)) {
+    //            int16_t out0, out1;
+    //            out0 = *ptr++;
+    //            out1 = (stereo ? *ptr++ : out0);
 
-                // output left channel
-                clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
+    //            // output left channel
+    //            clampedAdd(obuf[reverseStereo], (out0 * (int)vol_l) / Mixer::MaxVolume::MIXER);
 
-                // output right channel
-                clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
+    //            // output right channel
+    //            clampedAdd(obuf[reverseStereo ^ 1], (out1 * (int)vol_r) / Mixer::MaxVolume::MIXER);
 
-                obuf += 2;
-            }
-            return (obuf - ostart) / 2;
-        }
+    //            obuf += 2;
+    //        }
+    //        return (obuf - ostart) / 2;
+    //    }
 
-        virtual int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
-            return ST_SUCCESS;
-        }
-    };
+    //    virtual int drain(int16_t* obuf, uint32_t osamp, uint16_t vol) {
+    //        return ST_SUCCESS;
+    //    }
+    //};
 
 
     template<bool stereo, bool reverseStereo>
     RateConverter* makeRateConverter(uint32_t inrate, uint32_t outrate) {
         if (inrate != outrate) {
             if ((inrate % outrate) == 0 && (inrate < 65536)) {
-                return new SimpleRateConverter<stereo, reverseStereo>(inrate, outrate);
+                return new converters::SimpleRateConverter<stereo, reverseStereo>(inrate, outrate);
             }
             else {
-                return new LinearRateConverter<stereo, reverseStereo>(inrate, outrate);
+                return new converters::LinearRateConverter<stereo, reverseStereo>(inrate, outrate);
             }
         }
         else {
-            return new CopyRateConverter<stereo, reverseStereo>();
+            return new converters::CopyRateConverter<stereo, reverseStereo>();
         }
     }
